Split up food and pets into 2 modules
